requirements.txt
Django==1.6.1
MySQL-python==1.2.5
django-debug-toolbar==1.0.1
django-social-auth==0.7.28
httplib2==0.8
oauth2==1.5.211
python-openid==2.2.5
sqlparse==0.1.10
wsgiref==0.1.2
